fix(check)!: remove print/rewrite yaml
This feature was very rudimentary. Other formatters are way better and
differ from what this feature outputted.
Let it to the formatters that are good at its job.

This also allows for upgrading to serde_yaml 0.9 as
no yaml to string is needed anymore.

BREAKING CHANGE: check --print-yaml and --rewrite-yaml are gone